+++ components/AppBar/examples/AppBarWKWebViewLargeContentNoBugExample.m
@@ -53,7 +53,7 @@
 
     [self addChildViewController:_appBar.headerViewController];
 
-    _appBar.navigationBar.inkColor = [UIColor colorWithWhite:(CGFloat)0.9 alpha:(CGFloat)0.1];
+    _appBar.navigationBar.inkColor = [UIColor colorWithWhite:0.9f alpha:0.1f];
 
     self.colorScheme =
         [[MDCSemanticColorScheme alloc] initWithDefaults:MDCColorSchemeDefaultsMaterial201804];
