+++ components/Banner/tests/unit/Theming/MDCBannerViewThemingTests.m
@@ -19,9 +19,9 @@
 #import "MaterialContainerScheme.h"
 
 // The opacity value applied to text view.
-static CGFloat const kTextViewOpacity = (CGFloat)0.87;
+static CGFloat const kTextViewOpacity = 0.87f;
 // The opacity value applied to divider.
-static CGFloat const kDividerOpacity = (CGFloat)0.12;
+static CGFloat const kDividerOpacity = 0.12f;
 
 /**
 This class is used for creating a @UIWindow with customized size category.
